Move `/internal` packages into `/pkg`
