+++ integration/integration_test.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/cenkalti/backoff/v3"
 	"github.com/containous/maesh/integration/try"
-	"github.com/containous/maesh/internal/k8s"
+	"github.com/containous/maesh/pkg/k8s"
 	"github.com/containous/traefik/v2/pkg/config/dynamic"
 	"github.com/containous/traefik/v2/pkg/safe"
 	"github.com/go-check/check"
