+++ pkg/controller/controller.go
@@ -12,10 +12,10 @@ import (
 	"time"
 
 	"github.com/cenkalti/backoff/v3"
-	"github.com/containous/maesh/internal/k8s"
-	"github.com/containous/maesh/internal/providers/base"
-	"github.com/containous/maesh/internal/providers/kubernetes"
-	"github.com/containous/maesh/internal/providers/smi"
+	"github.com/containous/maesh/pkg/k8s"
+	"github.com/containous/maesh/pkg/providers/base"
+	"github.com/containous/maesh/pkg/providers/kubernetes"
+	"github.com/containous/maesh/pkg/providers/smi"
 	"github.com/containous/traefik/v2/pkg/config/dynamic"
 	"github.com/containous/traefik/v2/pkg/safe"
 	accessInformer "github.com/deislabs/smi-sdk-go/pkg/gen/client/access/informers/externalversions"
