+++ pkg/controller/handler.go
@@ -1,7 +1,7 @@
 package controller
 
 import (
-	"github.com/containous/maesh/internal/k8s"
+	"github.com/containous/maesh/pkg/k8s"
 	log "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 )
