+++ pkg/providers/base/base_test.go
@@ -3,7 +3,7 @@ package base
 import (
 	"testing"
 
-	"github.com/containous/maesh/internal/k8s"
+	"github.com/containous/maesh/pkg/k8s"
 
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
