+++ pkg/providers/kubernetes/fixtures/build_configuration_simple.yaml
