+++ pkg/providers/kubernetes/fixtures/build_configuration_simple_http_middlewares.yaml
