+++ pkg/providers/kubernetes/fixtures/build_service_missing_endpoints.yaml
