+++ pkg/providers/kubernetes/fixtures/build_service_simple.yaml
