+++ pkg/providers/smi/fixtures/build_configuration_http_service.yaml
