+++ pkg/providers/smi/fixtures/mock.yaml
