+++ pkg/providers/smi/provider.go
@@ -8,8 +8,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/containous/maesh/internal/k8s"
-	"github.com/containous/maesh/internal/providers/base"
+	"github.com/containous/maesh/pkg/k8s"
+	"github.com/containous/maesh/pkg/providers/base"
 	"github.com/containous/traefik/v2/pkg/config/dynamic"
 	access "github.com/deislabs/smi-sdk-go/pkg/apis/access/v1alpha1"
 	specs "github.com/deislabs/smi-sdk-go/pkg/apis/specs/v1alpha1"
